Tests/LibWeb/Text/expected/css/style-invalidation/viewport-resize-targeted-restyle.txt

viewport width: 50px
inherited font size: 50px
monospace recascaded font size: 50px
percentage font size: 75px
percentage em width: 75px
percentage line height: 75px
calc percentage line height: 67.5px
explicit inherit width: 50px
pseudo width: 50px
pseudo originating font size: 50px
pseudo inherited font size: 50px
full invalidations: 0
style recomputations bounded: true
inherited recomputations present: true
